task(android12_support-core_sdk): Extract static nested class LaunchPendingIntentFactory SDK-1111
